requirements.txt
requests>=2.28.0
mailjet-rest>=1.3.0
pyyaml>=6.0
